0-Input Technology, an Exoskeleton For The Mind
Goal: Democratize Intelligence To Promote Inclusion In Society. 
The Subjective Artificial Intelligence project is a groundbreaking initiative focused on redefining Human-Computer interaction and knowledge processing.  Subjective Artificial Intelligence has the potential to revolution the world in every single aspect of our lives. From our Subjective Semantizer driving us to a Post-Education Society where we say "The Knowledge Is Alive", to our Subjective Thermo-Currency turning a new Post-Scarcity era. Subjective Thermo-Currency is the first currency that is backed by the immutable laws of Physics and is able to render the concept of money either Crypto-Currencies or Fiat obsolete. For the first time made possible by using our Subjective framework design from a  Trans-Humanistic perspective approach. Explore the endless possibilities of our Virtual BodyModifications and the features that they enable for Humans enabling us to have better lives. Because Technology... Has to be Subjective.